feat: bring up LVGL landscape display

+++ Core/Src/ltdc.c
@@ -247,7 +247,7 @@ void MX_LTDC_Init(void)
   pLayerCfg.Alpha0 = 0;
   pLayerCfg.BlendingFactor1 = LTDC_BLENDING_FACTOR1_PAxCA;
   pLayerCfg.BlendingFactor2 = LTDC_BLENDING_FACTOR2_PAxCA;
-  pLayerCfg.FBStartAdress = LCD_BUFFER;
+  pLayerCfg.FBStartAdress = (uint32_t)(uintptr_t)LCD_BUFFER;
   pLayerCfg.ImageWidth = 240;
   pLayerCfg.ImageHeight = 320;
   pLayerCfg.Backcolor.Blue = 0;
